Took the functionality into app.tsx
