+++ src/pages/Calendar/Calendar.test.tsx
@@ -41,7 +41,10 @@ describe("Calendar page", () => {
   });
 
   it("should render the calender with the current date highlighted", () => {
-    customRender(<Calendar completedTasks={mockData} />);
+    // const changeDate = (value: Dayjs) => {
+    //   setDate(new Date(value.year(), value.month(), value.date()));
+    // };
+    customRender(<Calendar completedTasks={mockData} date={new Date()} />);
     const cal = screen.queryByTestId("calendarComponent");
     expect(cal).toBeInTheDocument();
     const today: number = new Date().getDate();
@@ -51,7 +54,7 @@ describe("Calendar page", () => {
   });
 
   it("should render the current date completed tasks only", async () => {
-    customRender(<Calendar completedTasks={mockData} />);
+    customRender(<Calendar completedTasks={mockData} date={new Date()} />);
     const currentDateCompletedTaskHeading = await screen.findByText(
       /5am wake up/i
     );
